// src/webapi/Types.ts
export namespace DynamicsWebApi {
// ...
    export interface Config {
        /**The id (id) to use when displaying the connection to Dynamics365 in the TreeView */
        id?: string;
        /**The friendly name (name) to use when displaying the connection to Dynamics365 in the TreeView */
        name?: string;
        /**The access token (accessToken) to use when connecting to Dynamics365 */
        accessToken?: string;
        /**The credentials (username) to use when connecting to Dynamics365 */
        username?: string;
        /**The credentials (password) to use when connecting to Dynamics365 */
        password?: string;
        /**The credentials (domain) to use when connecting to Dynamics365 */
        domain?: string;
        /**The credentials (workstation) to use when connecting to Dynamics365 */
        workstation?: string;
        /**A String representing a URL to Web API(webApiVersion not required if webApiUrl specified)[not used inside of CRM] */
        webApiUrl?: string;
        /**The version of Web API to use, for example: "8.1" */
        webApiVersion?: string;
        /**The ID of the organization */
        orgId?:string;
        /**The name of the organization */
        orgName?:string;
        /**A String representing the GUID value for the Dynamics 365 system user id.Impersonates the user. */
        impersonate?: string;
        /**A function that is called when a security token needs to be refreshed. */
        onTokenRefresh?: (callback: OnTokenAcquiredCallback) => void;
        /**Sets Prefer header with value "odata.include-annotations=" and the specified annotation.Annotations provide additional information about lookups, options sets and other complex attribute types.*/
        includeAnnotations?: string;
        /**Sets the odata.maxpagesize preference value to request the number of entities returned in the response. */
        maxPageSize?: string;
        /**Sets Prefer header request with value "return=representation".Use this property to return just created or updated entity in a single request.*/
        returnRepresentation?: boolean;
        /**Indicates whether to use Entity Logical Names instead of Collection Logical Names.*/
        useEntityNames?: boolean;
        /**Sets a number of milliseconds before a request times out */
        timeout?: number;
    }

    /** Callback with an acquired token called by DynamicsWebApi; "token" argument can be a string or an object with a property {accessToken: <token>}  */
    export interface OnTokenAcquiredCallback {
        (token: any): void;
    }
// ...
}